# src/main/resources/openapi/card-service.yaml
openapi: 3.0.3
info:
  title: Card Service API
  version: 1.0.0
  description: API for managing debit & credit cards
servers:
  - url: /api/v1
paths:
...
  /cards/{cardId}/movements:
    get:
      summary: Get card movements (last 10)
      description: Retrieves the last 10 movements for credit or debit card
      operationId: getCardMovements
      parameters:
        - name: cardId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: Card movements retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardMovementsResponse'
        '404':
          description: Card not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /cards/customers/{customerId}/cards-summary:
    get:
      summary: Get customer cards summary
      description: Returns all cards (credit and debit) for a specific customer
      operationId: getCustomerCardsSummary
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer cards summary retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerCardsSummaryResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /cards/reports/period:
    get:
      summary: Generate cards report for specific period
      description: Returns aggregated statistics for all card types in specified period
      operationId: generateCardsPeriodicReport
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Cards report generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CardsPeriodicReportResponse'
        '400':
          description: Invalid date range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    CardMovementsResponse:
      type: object
      required:
        - cardId
        - cardType
        - movements
        - totalCount
      properties:
        cardId:
          type: string
          example: "64f3cbb5f8a1e53d2f7c34e2"
        cardType:
          type: string
          enum: [ "CREDIT", "DEBIT" ]
          example: "DEBIT"
        movements:
          type: array
          items:
            $ref: '#/components/schemas/CardMovement'
        totalCount:
          type: integer
          example: 10
        retrievedAt:
          type: string
          format: date-time
    CardMovement:
      type: object
      properties:
        transactionId:
          type: string
          example: "txn_123456789"
        amount:
          type: number
          format: double
          example: 150.00
        transactionType:
          type: string
          enum: [ "PURCHASE", "WITHDRAWAL", "PAYMENT", "CHARGE", "AUTHORIZATION" ]
          example: "PURCHASE"
        merchantInfo:
          type: string
          example: "Amazon Purchase"
        accountsAffected:
          type: array
          items:
            type: object
            properties:
              accountId:
                type: string
              amountDeducted:
                type: number
                format: double
        processedAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [ "COMPLETED", "PENDING", "FAILED" ]
...
    CustomerCardsSummaryResponse:
      type: object
      required:
        - customerId
        - creditCards
        - debitCards
        - totalActiveCards
      properties:
        customerId:
          type: string
        creditCards:
          type: array
          items:
            $ref: '#/components/schemas/CreditCardSummary'
        debitCards:
          type: array
          items:
            $ref: '#/components/schemas/DebitCardSummary'
        totalActiveCards:
          type: integer
        retrievedAt:
          type: string
          format: date-time
    CreditCardSummary:
      type: object
      properties:
        cardId:
          type: string
        cardNumber:
          type: string
        availableCredit:
          type: number
        currentBalance:
          type: number
        isActive:
          type: boolean
    DebitCardSummary:
      type: object
      properties:
        cardId:
          type: string
        cardNumber:
          type: string
        primaryAccountId:
          type: string
        isActive:
          type: boolean
...
    CardsPeriodicReportResponse:
      type: object
      required:
        - reportPeriod
        - creditCards
        - debitCards
        - generatedAt
      properties:
        reportPeriod:
          type: object
          properties:
            startDate:
              type: string
              format: date
            endDate:
              type: string
              format: date
        creditCards:
          $ref: '#/components/schemas/CreditCardsReport'
        debitCards:
          $ref: '#/components/schemas/DebitCardsReport'
        generatedAt:
          type: string
          format: date-time
    CreditCardsReport:
      type: object
      properties:
        totalActiveCards:
          type: integer
        totalTransactions:
          type: integer
        totalAmountTransacted:
          type: number
          format: double
        averageCreditUtilization:
          type: number
          format: double
    DebitCardsReport:
      type: object
      properties:
        totalActiveCards:
          type: integer
        totalTransactions:
          type: integer
        totalAmountTransacted:
          type: number
          format: double
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "PERSON_ALREADY_HAS_CREDIT_CARD"
        message:
          type: string
          example: "Personal customers can only have one active credit card"
        timestamp:
          type: string
          format: date-time
          example: "2025-08-24T10:15:30Z"
      required:
        - code
        - message
        - timestamp